plugins/discord-connector/requirements.txt
discord.py-self>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
